Hw04 auth

+++ controllers/contacts.js
@@ -1,12 +1,24 @@
 const { Contact } = require("../models/contact");
-
 const { HttpError, ctrlWrapper } = require("../helpers");
 
 const listContacts = async (req, res) => {
-  const result = await Contact.find();
+  const { _id: owner } = req.user;
+  const { page = 1, limit = 10, favorite } = req.query;
+  const skip = (page - 1) * limit;
+
+  // we can pass the 2nd argument to find() to include fields ("password" or {password:1}) or exclude fields ("-password" or {password: 0})
+  const result = await Contact.find(
+    { owner, favorite },
+    "-createdAt -updatedAt",
+    {
+      limit,
+      skip,
+    }
+  ).populate("owner", "name email");
+  // populate() tells mongoose to get this "owner" field, find the collection it belongs to (using "ref" in Contact model), find the object with this ID and paste the whole object instead of this "owner" field
 
   if (!result) {
-    throw HttpError(404, "Not found");
+    throw HttpError(404);
   }
 
   res.status(200).json(result);
@@ -14,30 +26,49 @@ const listContacts = async (req, res) => {
 
 const getContactById = async (req, res) => {
   const { contactId } = req.params;
+  const { _id: owner } = req.user;
+
   const result = await Contact.findById(contactId);
 
+  // compares owner of token and owner of contact
+  if (result.owner.toString() !== owner.toString()) {
+    throw HttpError(401);
+  }
+
   if (!result) {
-    throw HttpError(404, "Not found");
+    throw HttpError(404);
   }
 
   res.status(200).json(result);
 };
 
 const addContact = async (req, res) => {
-  const result = await Contact.create(req.body);
-  if (!result) {
-    throw HttpError(404, "Not found");
+  const { _id: owner } = req.user; // we rename an "_id" to "owner" to match the User Schema
+  const newContact = await Contact.create({ ...req.body, owner });
+
+  // TODO: check if this contact already exists in the contacts of this user
+
+  if (!newContact) {
+    throw HttpError(422);
   }
 
-  res.status(201).json(result);
+  res.status(201).json(newContact);
 };
 
 const removeContact = async (req, res) => {
   const { contactId } = req.params;
+  const { _id: owner } = req.user;
+
+  // compares owner of token and owner of contact
+  const contactFromDB = await Contact.findById(contactId);
+  if (contactFromDB.owner.toString() !== owner.toString()) {
+    throw HttpError(401);
+  }
+
   const result = await Contact.findByIdAndRemove(contactId);
 
   if (!result) {
-    throw HttpError(404, "Not found");
+    throw HttpError(404);
   }
   // if we set res.status(204), the body of the message won't be sent in res.json()
   res.json("contact deleted");
@@ -45,16 +76,25 @@ const removeContact = async (req, res) => {
 
 const updateContact = async (req, res) => {
   const { contactId } = req.params;
+  const { _id: owner } = req.user;
 
   if (!req.body) {
-    throw HttpError(400, "Bad request");
+    throw HttpError(400);
   }
+
+  // compares owner of token and owner of contact
+  const contactFromDB = await Contact.findById(contactId);
+  if (contactFromDB.owner.toString() !== owner.toString()) {
+    throw HttpError(401);
+  }
+
   const result = await Contact.findByIdAndUpdate(contactId, req.body, {
     new: true,
+    select: "-createdAt",
   });
 
   if (!result) {
-    throw HttpError(404, "Not found");
+    throw HttpError(404);
   }
 
   res.status(200).json(result);
@@ -62,17 +102,25 @@ const updateContact = async (req, res) => {
 
 const updateStatusContact = async (req, res) => {
   const { contactId } = req.params;
+  const { _id: owner } = req.user;
 
   if (!req.body) {
     throw HttpError(400, "Missing field: favorite");
   }
 
+  // compares owner of token and owner of contact
+  const contactFromDB = await Contact.findById(contactId);
+  if (contactFromDB.owner.toString() !== owner.toString()) {
+    throw HttpError(401);
+  }
+
   const result = await Contact.findByIdAndUpdate(contactId, req.body, {
     new: true,
+    select: "-createdAt -updatedAt",
   });
 
   if (!result) {
-    throw HttpError(404, "Not found");
+    throw HttpError(404);
   }
 
   res.status(200).json(result);
